Moved scoped class into new files, and removed smart pointer wrapper
